refactor: add board, mailboxAdmin to wallet, move mailboxAdmin functionality to lib-mailbox
