+++ packages/cosmic-swingset/lib/ag-solo/vats/vat-wallet.js
@@ -47,9 +47,17 @@ function build(E, _D, _log) {
   const { publish: pursesPublish, subscribe: purseSubscribe } = pubsub(E);
   const { publish: inboxPublish, subscribe: inboxSubscribe } = pubsub(E);
 
-  async function startup(zoe, registry) {
-    wallet = await makeWallet(E, zoe, registry, pursesPublish, inboxPublish);
+  async function startup({ zoe, registry, board, mailboxAdmin }) {
+    wallet = await makeWallet({
+      zoe,
+      mailboxAdmin,
+      board,
+      registry,
+      pursesStateChangeHandler: pursesPublish,
+      inboxStateChangeHandler: inboxPublish,
+    });
   }
+
 
   async function getWallet() {
     return harden(wallet);
