+++ packages/cosmic-swingset/test/unitTests/test-lib-wallet.js
@@ -10,6 +10,8 @@ import harden from '@agoric/harden';
 import { makeGetInstanceHandle } from '@agoric/zoe/src/clientSupport';
 
 import { makeWallet } from '../../lib/ag-solo/vats/lib-wallet';
+import { makeBoard } from '../../lib/ag-solo/vats/lib-board';
+import { makeMailboxAdmin } from '../../lib/ag-solo/vats/lib-mailbox';
 
 const setupTest = async () => {
   const contractRoot = require.resolve(
@@ -29,6 +31,8 @@ const setupTest = async () => {
   const rpgBundle = produceIssuer('rpg', 'strSet');
   const zoe = makeZoe({ require });
   const registry = makeRegistrar();
+  const board = makeBoard();
+  const mailboxAdmin = makeMailboxAdmin(board);
 
   const installationHandle = zoe.install(source, moduleFormat);
 
@@ -45,18 +49,21 @@ const setupTest = async () => {
     instanceHandle,
   );
 
-  const wallet = await makeWallet(
-    E,
+  const wallet = await makeWallet({
     zoe,
     registry,
+    board,
+    mailboxAdmin,
     pursesStateChangeHandler,
     inboxStateChangeHandler,
-  );
+  });
   return {
     moolaBundle,
     rpgBundle,
     zoe,
     registry,
+    mailboxAdmin,
+    board,
     wallet,
     invite,
     installationHandle,
